+++ kinto_wizard/kinto2yaml.py
@@ -6,10 +6,11 @@ def _sorted_principals(permissions):
     return {perm: sorted(principals) for perm, principals in permissions.items()}
 
 
-def introspect_server(client, bucket=None, collection=None, full=False):
+def introspect_server(client, bucket=None, collection=None, data=False, records=False):
     if bucket:
         logger.info("Only inspect bucket `{}`.".format(bucket))
-        bucket_info = introspect_bucket(client, bucket, collection=collection, full=full)
+        bucket_info = introspect_bucket(client, bucket, collection=collection,
+                                        data=data, records=records)
         if bucket_info:
             return {bucket: bucket_info}
         return {}
@@ -17,12 +18,13 @@ def introspect_server(client, bucket=None, collection=None, full=False):
     logger.info("Fetch buckets list.")
     buckets = client.get_buckets()
     return {
-        bucket['id']: introspect_bucket(client, bucket['id'], collection=collection, full=full)
+        bucket['id']: introspect_bucket(client, bucket['id'], collection=collection,
+                                        data=data, records=records)
         for bucket in buckets
     }
 
 
-def introspect_bucket(client, bid, collection=None, full=False):
+def introspect_bucket(client, bid, collection=None, data=False, records=False):
     logger.info("Fetch information of bucket {!r}".format(bid))
     try:
         bucket = client.get_bucket(id=bid)
@@ -37,7 +39,8 @@ def introspect_bucket(client, bid, collection=None, full=False):
     if collection:
         result = {
             'permissions': _sorted_principals(permissions),
-            'collections': {collection: introspect_collection(client, bid, collection, full=full)}
+            'collections': {collection: introspect_collection(client, bid, collection,
+                                                              data=data, records=records)}
         }
     else:
         collections = client.get_collections(bucket=bid)
@@ -45,29 +48,30 @@ def introspect_bucket(client, bid, collection=None, full=False):
         result = {
             'permissions': _sorted_principals(permissions),
             'collections': {
-                collection['id']: introspect_collection(client, bid, collection['id'], full=full)
+                collection['id']: introspect_collection(client, bid, collection['id'],
+                                                        data=data, records=records)
                 for collection in collections
             },
             'groups': {
-                group['id']: introspect_group(client, bid, group['id'], full=full)
+                group['id']: introspect_group(client, bid, group['id'], data=data)
                 for group in groups
             }
         }
-    if full:
+    if data:
         result['data'] = bucket['data']
     return result
 
 
-def introspect_collection(client, bid, cid, full=False):
+def introspect_collection(client, bid, cid, data=False, records=False):
     logger.info("Fetch information of collection {!r}/{!r}".format(bid, cid))
     collection = client.get_collection(bucket=bid, id=cid)
     result = {
         'permissions': _sorted_principals(collection['permissions']),
     }
-    if full:
+    if data:
         result['data'] = collection['data']
 
-        # If full, include records.
+    if records:
         records = client.get_records(bucket=bid, collection=cid)
         result['records'] = {
             # XXX: we don't show permissions, until we have a way to fetch records
@@ -77,13 +81,13 @@ def introspect_collection(client, bid, cid, full=False):
     return result
 
 
-def introspect_group(client, bid, gid, full=False):
+def introspect_group(client, bid, gid, data=False):
     logger.info("Fetch information of group {!r}/{!r}".format(bid, gid))
     group = client.get_group(bucket=bid, id=gid)
     result = {
         'permissions': _sorted_principals(group['permissions'])
     }
-    data = group['data'] if full else {}
+    data = group['data'] if data else {}
     data['members'] = sorted(group['data']['members'])
     result['data'] = data
     return result